# internal/api/v1/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Tailnet API",
    "title": "Tailnet API",
    "termsOfService": "https://github.com/sudosu404/providers/blob/main/LICENSE",
    "contact": {
      "name": "Yusing",
      "url": "https://github.com/sudosu404/providers/issues"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/sudosu404/providers/blob/main/LICENSE"
    },
    "version": "1.0"
  },
  "basePath": "/api/v1",
  "paths": {
...
    "/file/validate": {
      "post": {
        "description": "Validate file",
        "consumes": [
          "text/plain"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "file"
        ],
        "summary": "Validate file",
        "parameters": [
          {
            "enum": [
              "config",
              "provider",
              "middleware"
            ],
            "type": "string",
            "description": "Type",
            "name": "type",
            "in": "query",
            "required": true
          },
          {
            "description": "File content",
            "name": "file",
            "in": "body",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "File validated",
            "schema": {
              "$ref": "#/definitions/SuccessResponse"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "417": {
            "description": "Validation failed",
            "schema": {}
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "x-id": "validate",
        "operationId": "validate"
      }
    },
...
    "/homepage/set/category_order": {
      "post": {
        "description": "Set homepage category order.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "homepage"
        ],
        "summary": "Set homepage category order",
        "parameters": [
          {
            "description": "Override category order",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/HomepageOverrideCategoryOrderParams"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SuccessResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "x-id": "set-category-order",
        "operationId": "set-category-order"
      }
    },
    "/homepage/set/item": {
      "post": {
        "description": "Override single homepage item.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "homepage"
        ],
        "summary": "Override single homepage item",
        "parameters": [
          {
            "description": "Override single item",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/HomepageOverrideItemParams"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SuccessResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "x-id": "set-item",
        "operationId": "set-item"
      }
    },
    "/homepage/set/item_all_sort_order": {
      "post": {
        "description": "Set homepage item all sort order.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "homepage"
        ],
        "summary": "Set homepage item all sort order",
        "parameters": [
          {
            "description": "Set item all sort order",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/HomepageOverrideItemAllSortOrderParams"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SuccessResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "x-id": "set-item-all-sort-order",
        "operationId": "set-item-all-sort-order"
      }
    },
    "/homepage/set/item_fav_sort_order": {
      "post": {
        "description": "Set homepage item fav sort order.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "homepage"
        ],
        "summary": "Set homepage item fav sort order",
        "parameters": [
          {
            "description": "Set item fav sort order",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/HomepageOverrideItemFavSortOrderParams"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SuccessResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "x-id": "set-item-fav-sort-order",
        "operationId": "set-item-fav-sort-order"
      }
    },
    "/homepage/set/item_favorite": {
      "post": {
        "description": "Set homepage item favorite.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "homepage"
        ],
        "summary": "Set homepage item favorite",
        "parameters": [
          {
            "description": "Set item favorite",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/HomepageOverrideItemFavoriteParams"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SuccessResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "x-id": "set-item-favorite",
        "operationId": "set-item-favorite"
      }
    },
    "/homepage/set/item_sort_order": {
      "post": {
        "description": "Set homepage item sort order.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "homepage"
        ],
        "summary": "Set homepage item sort order",
        "parameters": [
          {
            "description": "Set item sort order",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/HomepageOverrideItemSortOrderParams"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SuccessResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "x-id": "set-item-sort-order",
        "operationId": "set-item-sort-order"
      }
    },
    "/homepage/set/item_visible": {
      "post": {
        "description": "POST list of item ids and visibility value.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "homepage"
        ],
        "summary": "Set homepage item visibility",
        "parameters": [
          {
            "description": "Set item visibility",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/HomepageOverrideItemVisibleParams"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SuccessResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "x-id": "set-item-visible",
        "operationId": "set-item-visible"
      }
    },
    "/homepage/set/items_batch": {
      "post": {
        "description": "Override multiple homepage items.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "homepage"
        ],
        "summary": "Override multiple homepage items",
        "parameters": [
          {
            "description": "Override multiple items",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/HomepageOverrideItemsBatchParams"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SuccessResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "x-id": "set-items-batch",
        "operationId": "set-items-batch"
      }
    },
...
    "/route/providers": {
      "get": {
        "description": "List route providers",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "route",
          "websocket"
        ],
        "summary": "List route providers",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/RouteProvider"
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "x-id": "providers",
        "operationId": "providers"
      }
    },
    "/route/{which}": {
      "get": {
        "description": "List route",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "route"
        ],
        "summary": "List route",
        "parameters": [
          {
            "type": "string",
            "description": "Route name",
            "name": "which",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Route"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "x-id": "route",
        "operationId": "route"
      }
    },
...
  },
  "definitions": {
...
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string",
          "x-nullable": true
        },
        "message": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        }
      },
      "x-nullable": false,
      "x-omitempty": false
    },
...
    "HomepageItemConfig": {
      "type": "object",
      "properties": {
        "category": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "description": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "favorite": {
          "type": "boolean",
          "x-nullable": false,
          "x-omitempty": false
        },
        "icon": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "name": {
          "description": "display name",
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "show": {
          "type": "boolean",
          "x-nullable": false,
          "x-omitempty": false
        },
        "url": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "widget_config": {
          "allOf": [
            {
              "$ref": "#/definitions/widgets.Config"
            }
          ],
          "x-nullable": true
        }
      },
      "x-nullable": false,
      "x-omitempty": false
    },
...
    "HomepageOverrideCategoryOrderParams": {
      "type": "object",
      "properties": {
        "value": {
          "type": "integer",
          "x-nullable": false,
          "x-omitempty": false
        },
        "which": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        }
      },
      "x-nullable": false,
      "x-omitempty": false
    },
    "HomepageOverrideItemAllSortOrderParams": {
      "type": "object",
      "properties": {
        "value": {
          "type": "integer",
          "x-nullable": false,
          "x-omitempty": false
        },
        "which": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        }
      },
      "x-nullable": false,
      "x-omitempty": false
    },
    "HomepageOverrideItemFavSortOrderParams": {
      "type": "object",
      "properties": {
        "value": {
          "type": "integer",
          "x-nullable": false,
          "x-omitempty": false
        },
        "which": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        }
      },
      "x-nullable": false,
      "x-omitempty": false
    },
    "HomepageOverrideItemFavoriteParams": {
      "type": "object",
      "properties": {
        "value": {
          "type": "boolean",
          "x-nullable": false,
          "x-omitempty": false
        },
        "which": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-nullable": false,
          "x-omitempty": false
        }
      },
      "x-nullable": false,
      "x-omitempty": false
    },
    "HomepageOverrideItemParams": {
      "type": "object",
      "properties": {
        "value": {
          "$ref": "#/definitions/HomepageItemConfig",
          "x-nullable": false,
          "x-omitempty": false
        },
        "which": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        }
      },
      "x-nullable": false,
      "x-omitempty": false
    },
    "HomepageOverrideItemSortOrderParams": {
      "type": "object",
      "properties": {
        "value": {
          "type": "integer",
          "x-nullable": false,
          "x-omitempty": false
        },
        "which": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        }
      },
      "x-nullable": false,
      "x-omitempty": false
    },
    "HomepageOverrideItemVisibleParams": {
      "type": "object",
      "properties": {
        "value": {
          "type": "boolean",
          "x-nullable": false,
          "x-omitempty": false
        },
        "which": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-nullable": false,
          "x-omitempty": false
        }
      },
      "x-nullable": false,
      "x-omitempty": false
    },
    "HomepageOverrideItemsBatchParams": {
      "type": "object",
      "properties": {
        "value": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/HomepageItemConfig"
          },
          "x-nullable": false,
          "x-omitempty": false
        }
      },
      "x-nullable": false,
      "x-omitempty": false
    },
...
    "RequestLoggerConfig": {
      "type": "object",
      "properties": {
        "buffer_size": {
          "description": "Deprecated: buffer size is adjusted dynamically",
          "type": "integer",
          "x-nullable": false,
          "x-omitempty": false
        },
        "fields": {
          "$ref": "#/definitions/accesslog.Fields",
          "x-nullable": false,
          "x-omitempty": false
        },
        "filters": {
          "$ref": "#/definitions/accesslog.Filters",
          "x-nullable": false,
          "x-omitempty": false
        },
        "format": {
          "type": "string",
          "enum": [
            "common",
            "combined",
            "json"
          ],
          "x-nullable": false,
          "x-omitempty": false
        },
        "path": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "retention": {
          "$ref": "#/definitions/LogRetention",
          "x-nullable": false,
          "x-omitempty": false
        },
        "rotate_interval": {
          "type": "integer",
          "x-nullable": false,
          "x-omitempty": false
        },
        "stdout": {
          "type": "boolean",
          "x-nullable": false,
          "x-omitempty": false
        }
      },
      "x-nullable": false,
      "x-omitempty": false
    },
    "Route": {
      "type": "object",
      "properties": {
        "access_log": {
          "allOf": [
            {
              "$ref": "#/definitions/RequestLoggerConfig"
            }
          ],
          "x-nullable": true
        },
        "agent": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "alias": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "container": {
          "description": "Docker only",
          "allOf": [
            {
              "$ref": "#/definitions/Container"
            }
          ],
          "x-nullable": true
        },
        "disable_compression": {
          "type": "boolean",
          "x-nullable": false,
          "x-omitempty": false
        },
        "excluded": {
          "type": "boolean",
          "x-nullable": true
        },
        "excluded_reason": {
          "type": "string",
          "x-nullable": true
        },
        "health": {
          "description": "for swagger",
          "allOf": [
            {
              "$ref": "#/definitions/HealthJSON"
            }
          ],
          "x-nullable": false,
          "x-omitempty": false
        },
        "healthcheck": {
          "$ref": "#/definitions/HealthCheckConfig",
          "x-nullable": false,
          "x-omitempty": false
        },
        "homepage": {
          "$ref": "#/definitions/HomepageItemConfig",
          "x-nullable": false,
          "x-omitempty": false
        },
        "host": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "idlewatcher": {
          "allOf": [
            {
              "$ref": "#/definitions/IdlewatcherConfig"
            }
          ],
          "x-nullable": true
        },
        "load_balance": {
          "allOf": [
            {
              "$ref": "#/definitions/LoadBalancerConfig"
            }
          ],
          "x-nullable": true
        },
        "lurl": {
          "description": "private fields",
          "type": "string",
          "x-nullable": true
        },
        "middlewares": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/types.LabelMap"
          },
          "x-nullable": true
        },
        "no_tls_verify": {
          "type": "boolean",
          "x-nullable": false,
          "x-omitempty": false
        },
        "path_patterns": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-nullable": true
        },
        "port": {
          "$ref": "#/definitions/Port",
          "x-nullable": false,
          "x-omitempty": false
        },
        "provider": {
          "description": "for backward compatibility",
          "type": "string",
          "x-nullable": true
        },
        "purl": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "response_header_timeout": {
          "type": "integer",
          "x-nullable": false,
          "x-omitempty": false
        },
        "root": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "rule_file": {
          "type": "string",
          "x-nullable": true
        },
        "rules": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/rules.Rule"
          },
          "x-nullable": false,
          "x-omitempty": false
        },
        "scheme": {
          "type": "string",
          "enum": [
            "http",
            "https",
            "tcp",
            "udp",
            "fileserver"
          ],
          "x-nullable": false,
          "x-omitempty": false
        },
        "ssl_certificate": {
          "description": "Path to client certificate",
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "ssl_certificate_key": {
          "description": "Path to client certificate key",
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "ssl_protocols": {
          "description": "Allowed TLS protocols",
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-nullable": false,
          "x-omitempty": false
        },
        "ssl_server_name": {
          "description": "SSL/TLS proxy options (nginx-like)",
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "ssl_trusted_certificate": {
          "description": "Path to trusted CA certificates",
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        }
      },
      "x-nullable": false,
      "x-omitempty": false
    },
    "RouteProvider": {
      "type": "object",
      "properties": {
        "full_name": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "short_name": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        }
      },
      "x-nullable": false,
      "x-omitempty": false
    },
...
    "SuccessResponse": {
      "type": "object",
      "properties": {
        "details": {
          "type": "object",
          "additionalProperties": {},
          "x-nullable": true
        },
        "message": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        }
      },
      "x-nullable": false,
      "x-omitempty": false
    },
...
  },
...
}